[supervision] Checking collector key at host level.
On the host running the agent when PostgreSQL is not started,
probes do not return any information related to PG instance then
when the collector's checking the agent key, we don't have any pg_port
and pg_data to find the right instance entry in the repository.
It that case, collector will check the keys of all instances entries with
the same hostname.