Extract statements's metadata into a separate table
We create a "statement" table to store metadata received from the agent
by request on the /statements endpoint in a new "statements" table. The
statements_history_current now references the multi-column primary key
of this table.

This should help limiting the storage volume by avoiding to store the
same query multiple times and makes it easier to add indexes.