Build RPM with Python 3 on CentOS 7
We now always build the RPM package with Python 3, thus removing conditions on
rhel values in .spec file (notice that packages names are the same). We now use
the __python3 macro (pulled from python3-rpm-macros package, thus added to
BuildRequires). Additionally, we remove python_sitelib macro that
would be wrong now and use python3_sitelib.

Since there would be no "upstream" version, we bump the pkgrevision of RPM
packaging to ship this change and document it in changelog.

As a consequence of building a python3 rpm package, we now run func tests with
python3 on centos7 in CI and require that python3.6 unit tests pass (instead of
python2.7 ones previously) by consistency.