Improve error handling in agent
Error raised when applying plugin where not shown in log and even
when a 500 error, 200 where shown.
We avoid using after request hooks and add log in wsgi() surcharge.
We removed ErrorPlugin to handle errors in default_error_handler().
We add an /error entry point on the agent to test for error generation.