Get the scheduled reindex from schema endpoint
Previously we were getting the scheduled reindexes only for the table. So we didn't get the scheduled reindexes for the indexes since they are not provided by the agent/table/*/reindex/scheduled.
However this requires the list to be filtered if we then want the list of scheduled reindexes for the table only.