Don't rely on data to show 'no data' in place of sparkline
Using v-if here generates the following error:
Unhandled error during execution of scheduler flush

This is apparently due to the fact that in some cases the data changes from
'valid data' -> null. Because the sparkline was already rendered, the content
of the 'root' element was changed. In this case, the element where to display
the 'no data' message being nested in the root element destroyed. The problem
is that Vue doesn't know that this element doesn't exist anymore and tries to
display it again.