Load jQuery and bootstrap from npm packages
We don't load jquery with <script> tag in the base templates.

jQuery built library is still used in the configuration and settings.notifications pages.
So we specifically load it using a <script> tag for those.

Changes to the $.ajax API have been made since we upgraded from 2.x to 3.x.