docs: Move mkdocs to top level directory
mkdocs fails to hot reload if configuration is in docs dir.

Fixes:

ERROR   -  Config value 'docs_dir': The 'docs_dir' should not be the parent directory of the config file. Use a child directory instead so that the 'docs_dir' is a sibling of the config file.
Aborted with a configuration error!
ERROR   -  An error happened during the rebuild. The server will appear stuck until build errors are resolved.